+++ packages/cli/src/commands/sandbox/sandbox-secret/sandbox_secret_command_factory.ts
@@ -1,7 +1,7 @@
 import { CommandModule } from 'yargs';
 
-import { LocalAppNameResolver } from '../../../backend-identifier/local_app_name_resolver.js';
-import { SandboxIdResolver } from '../sandbox_id_resolver.js';
+import { LocalNamespaceResolver } from '../../../backend-identifier/local_namespace_resolver.js';
+import { SandboxBackendIdResolver } from '../sandbox_id_resolver.js';
 import { CwdPackageJsonLoader } from '../../../cwd_package_json_loader.js';
 import { SandboxSecretCommand } from './sandbox_secret_command.js';
 import { getSecretClient } from '@aws-amplify/backend-secret';
@@ -14,8 +14,8 @@ import { SandboxSecretListCommand } from './sandbox_secret_list_command.js';
  * Creates sandbox secret commands.
  */
 export const createSandboxSecretCommand = (): CommandModule => {
-  const sandboxIdResolver = new SandboxIdResolver(
-    new LocalAppNameResolver(new CwdPackageJsonLoader())
+  const sandboxIdResolver = new SandboxBackendIdResolver(
+    new LocalNamespaceResolver(new CwdPackageJsonLoader())
   );
 
   const secretClient = getSecretClient();