+++ packages/cli/src/commands/sandbox/sandbox_command_factory.ts
@@ -2,13 +2,12 @@ import { CommandModule } from 'yargs';
 import { SandboxCommand, SandboxCommandOptions } from './sandbox_command.js';
 import { SandboxSingletonFactory } from '@aws-amplify/sandbox';
 import { SandboxDeleteCommand } from './sandbox-delete/sandbox_delete_command.js';
-import { SandboxIdResolver } from './sandbox_id_resolver.js';
+import { SandboxBackendIdResolver } from './sandbox_id_resolver.js';
 import { CwdPackageJsonLoader } from '../../cwd_package_json_loader.js';
 import { ClientConfigGeneratorAdapter } from '../../client-config/client_config_generator_adapter.js';
 import { fromNodeProviderChain } from '@aws-sdk/credential-providers';
-import { LocalAppNameResolver } from '../../backend-identifier/local_app_name_resolver.js';
+import { LocalNamespaceResolver } from '../../backend-identifier/local_namespace_resolver.js';
 import { createSandboxSecretCommand } from './sandbox-secret/sandbox_secret_command_factory.js';
-import { SandboxBackendIdentifier } from '@aws-amplify/platform-core';
 import { SandboxEventHandlerFactory } from './sandbox_event_handler_factory.js';
 import { CommandMiddleware } from '../../command_middleware.js';
 
@@ -20,20 +19,35 @@ export const createSandboxCommand = (): CommandModule<
   SandboxCommandOptions
 > => {
   const credentialProvider = fromNodeProviderChain();
-  const sandboxIdResolver = new SandboxIdResolver(
-    new LocalAppNameResolver(new CwdPackageJsonLoader())
+  const sandboxBackendIdPartsResolver = new SandboxBackendIdResolver(
+    new LocalNamespaceResolver(new CwdPackageJsonLoader())
   );
-  const sandboxFactory = new SandboxSingletonFactory(sandboxIdResolver.resolve);
+
+  /**
+   * This callback allows sandbox to resolve the backend id using the name specified by the --name arg if present
+   * Otherwise, the default sandboxBackendIdPartsResolver.resolve() value is used
+   * @param sandboxName A customer specified name to use as part of the sandbox identifier (the --name arg to sandbox)
+   */
+  const sandboxBackendIdentifierResolver = async (sandboxName?: string) => {
+    const sandboxBackendIdParts = await sandboxBackendIdPartsResolver.resolve();
+    if (!sandboxName) {
+      return sandboxBackendIdParts;
+    }
+    return {
+      ...sandboxBackendIdParts,
+      name: sandboxName,
+    };
+  };
+
+  const sandboxFactory = new SandboxSingletonFactory(
+    sandboxBackendIdentifierResolver
+  );
   const clientConfigGeneratorAdapter = new ClientConfigGeneratorAdapter(
     credentialProvider
   );
-  const getBackendIdentifier = async (appName?: string) => {
-    const sandboxId = appName ?? (await sandboxIdResolver.resolve());
-    return new SandboxBackendIdentifier(sandboxId);
-  };
 
   const eventHandlerFactory = new SandboxEventHandlerFactory(
-    getBackendIdentifier
+    sandboxBackendIdentifierResolver
   );
 
   const commandMiddleWare = new CommandMiddleware();