+++ packages/create-amplify/src/amplify_project_creator.ts
@@ -16,8 +16,8 @@ export class AmplifyProjectCreator {
   ];
 
   // TODO after API-Next is GA change to: `aws-amplify`
-  // https://github.com/aws-amplify/samsara-cli/issues/332
-  private readonly defaultProdPackages = ['aws-amplify@api-v6-models'];
+  // https://github.com/aws-amplify/amplify-backend/issues/332
+  private readonly defaultProdPackages = ['aws-amplify@unstable'];
 
   /**
    * Orchestrator for the create-amplify workflow.