+++ packages/deployed-backend-client/src/backend_output_fetcher_factory.ts
@@ -1,6 +1,6 @@
 import { AmplifyClient } from '@aws-sdk/client-amplify';
 import { CloudFormationClient } from '@aws-sdk/client-cloudformation';
-import { BackendIdentifier } from './backend_identifier.js';
+import { DeployedBackendIdentifier } from './deployed_backend_identifier.js';
 import {
   AppNameAndBranchBackendIdentifier,
   AppNameAndBranchMainStackNameResolver,
@@ -9,25 +9,27 @@ import {
   PassThroughMainStackNameResolver,
   StackIdentifier,
 } from './stack-name-resolvers/passthrough_main_stack_name_resolver.js';
-import { UniqueBackendIdentifierMainStackNameResolver } from './stack-name-resolvers/unique_deployment_identifier_main_stack_name_resolver.js';
+import { BackendIdentifierMainStackNameResolver } from './stack-name-resolvers/backend_identifier_main_stack_name_resolver.js';
 import { StackMetadataBackendOutputRetrievalStrategy } from './stack_metadata_output_retrieval_strategy.js';
-import { UniqueBackendIdentifier } from '@aws-amplify/plugin-types';
+import { BackendIdentifier } from '@aws-amplify/plugin-types';
 
 /**
- * Asserts that a BackendIdentifier is a UniqueBackendIdentifier
+ * Asserts that a BackendIdentifier is a BackendIdentifier
  */
-export const isUniqueBackendIdentifier = (
-  backendIdentifier: BackendIdentifier
-): backendIdentifier is UniqueBackendIdentifier => {
+export const isBackendIdentifier = (
+  backendIdentifier: DeployedBackendIdentifier
+): backendIdentifier is BackendIdentifier => {
   return (
-    'backendId' in backendIdentifier && 'disambiguator' in backendIdentifier
+    'namespace' in backendIdentifier &&
+    'name' in backendIdentifier &&
+    'type' in backendIdentifier
   );
 };
 /**
  * Asserts that a BackendIdentifier is a AppNameAndBranchBackendIdentifier
  */
 export const isAppNameAndBranchIdentifier = (
-  backendIdentifier: BackendIdentifier
+  backendIdentifier: DeployedBackendIdentifier
 ): backendIdentifier is AppNameAndBranchBackendIdentifier => {
   return 'appName' in backendIdentifier;
 };
@@ -35,7 +37,7 @@ export const isAppNameAndBranchIdentifier = (
  * Asserts that a BackendIdentifier is a StackIdentifier
  */
 export const isStackIdentifier = (
-  backendIdentifier: BackendIdentifier
+  backendIdentifier: DeployedBackendIdentifier
 ): backendIdentifier is StackIdentifier => {
   return 'stackName' in backendIdentifier;
 };
@@ -50,16 +52,16 @@ export class BackendOutputFetcherFactory {
     private cfnClient: CloudFormationClient,
     private amplifyClient: AmplifyClient
   ) {}
-  getStrategy = (backendIdentifier: BackendIdentifier) => {
+  getStrategy = (backendIdentifier: DeployedBackendIdentifier) => {
     if (isStackIdentifier(backendIdentifier)) {
       return new StackMetadataBackendOutputRetrievalStrategy(
         this.cfnClient,
         new PassThroughMainStackNameResolver(backendIdentifier)
       );
-    } else if (isUniqueBackendIdentifier(backendIdentifier)) {
+    } else if (isBackendIdentifier(backendIdentifier)) {
       return new StackMetadataBackendOutputRetrievalStrategy(
         this.cfnClient,
-        new UniqueBackendIdentifierMainStackNameResolver(backendIdentifier)
+        new BackendIdentifierMainStackNameResolver(backendIdentifier)
       );
     }
     return new StackMetadataBackendOutputRetrievalStrategy(