+++ packages/platform-core/src/cdk_context_key.ts
@@ -1,8 +1,11 @@
 /**
  * Keys for values that are passed between the Amplify deployment engine and the CDK synth process via CDK Context
  *
- * This enum does not currently capture all context keys. We can refactor incrementally as we need to use these keys in multiple places
+ * !!!WARNING!!!
+ * Changing the values of these keys could be a breaking change
  */
 export enum CDKContextKey {
-  DEPLOYMENT_TYPE = 'deployment-type',
+  BACKEND_NAME = 'amplify-backend-name',
+  BACKEND_NAMESPACE = 'amplify-backend-namespace',
+  DEPLOYMENT_TYPE = 'amplify-backend-type',
 }