+++ packages/sandbox/src/file_watching_sandbox.ts
@@ -2,6 +2,7 @@ import debounce from 'debounce-promise';
 import parcelWatcher, { subscribe } from '@parcel/watcher';
 import { AmplifySandboxExecutor } from './sandbox_executor.js';
 import {
+  BackendIdSandboxResolver,
   Sandbox,
   SandboxDeleteOptions,
   SandboxEvents,
@@ -16,7 +17,6 @@ import {
   CloudFormationClient,
   DescribeStacksCommand,
 } from '@aws-sdk/client-cloudformation';
-import { SandboxBackendIdentifier } from '@aws-amplify/platform-core';
 import { AmplifyPrompter } from '@aws-amplify/cli-core';
 import {
   FilesChangesTracker,
@@ -27,7 +27,7 @@ export const CDK_BOOTSTRAP_STACK_NAME = 'CDKToolkit';
 export const CDK_BOOTSTRAP_VERSION_KEY = 'BootstrapVersion';
 export const CDK_MIN_BOOTSTRAP_VERSION = 6;
 
-// TODO: finalize bootstrap url: https://github.com/aws-amplify/samsara-cli/issues/338
+// TODO: finalize bootstrap url: https://github.com/aws-amplify/amplify-backend/issues/338
 /**
  * Constructs Amplify Console bootstrap URL for a given region
  * @param region AWS region
@@ -48,7 +48,7 @@ export class FileWatchingSandbox extends EventEmitter implements Sandbox {
    * Creates a watcher process for this instance
    */
   constructor(
-    private readonly sandboxId: string,
+    private readonly backendIdSandboxResolver: BackendIdSandboxResolver,
     private readonly executor: AmplifySandboxExecutor,
     private readonly cfnClient: CloudFormationClient,
     private readonly open = _open
@@ -179,11 +179,12 @@ export class FileWatchingSandbox extends EventEmitter implements Sandbox {
    * @inheritdoc
    */
   delete = async (options: SandboxDeleteOptions) => {
-    const sandboxAppId = options.name ?? this.sandboxId;
     console.log(
       '[Sandbox] Deleting all the resources in the sandbox environment...'
     );
-    await this.executor.destroy(new SandboxBackendIdentifier(sandboxAppId));
+    await this.executor.destroy(
+      await this.backendIdSandboxResolver(options.name)
+    );
     this.emit('successfulDeletion');
     console.log('[Sandbox] Finished deleting.');
   };
@@ -201,10 +202,9 @@ export class FileWatchingSandbox extends EventEmitter implements Sandbox {
   };
 
   private deploy = async (options: SandboxOptions) => {
-    const sandboxAppId = options.name ?? this.sandboxId;
     try {
       await this.executor.deploy(
-        new SandboxBackendIdentifier(sandboxAppId),
+        await this.backendIdSandboxResolver(options.name),
         // It's important to pass this as callback so that debounce does
         // not reset tracker prematurely
         this.shouldValidateAppSources