+++ packages/backend-output-schemas/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/backend-output-schemas
 
+## 0.3.0
+
+### Minor Changes
+
+- 71a63a16: Change stack naming strategy to include deployment type as a suffix
+
 ## 0.2.1
 
 ### Patch Changes